fakewifi: switch to x_catawampus_tr098_1_0
The immediate need was for OperatingFrequencyBand, but while
there we implemented support for all of the various
parameters in the extension data model.

Also add a test for all of the Wifi data models to ensure that
they point to the extension. We test for this by looking for
'OperatingFrequencyBand' as an exported parameter.

Change-Id: I26e4b493857648059d2d36903797d6cb1f36a978