cwmpd: stdout was unbuffered instead of line buffered as intended.
The primary upshot of this was that 'print "x"' would come out as two write()
calls: "x" and "\n".  If we had a subprogram running, such as ginstall, and
sharing the same logos, it would be possible for the subprogram to interject
its output between the two, so you'd get lines jumbled together sometimes.

Using line buffering, reasonably-sized lines will now be written as a single
atomic write(), reducing jumbling.

Change-Id: I49d45aba7ca310c187a19b67f5d3f8b21ad765a9