UiChoice to describe which TV UI will run.
ACS can set the (existing) UiType variable to "sage" or
"oregano", and can read the UiType variable back to see
what it is set to. We have UiType read back 'Unknown' if
not set, so that it is possible to tell whether it has
been set or not.

Nonetheless, if the UiType is unset the platform will choose
a default, and we often want to know what UI it is going to
choose. Provide a new UiChoice parameter to tell what the
choice will be even if UiType is not set.

Fixes b/19652357

Change-Id: I83eb96be8c797817664ce253695497c77e484513