cwmplog: Truncate GetParameterValues & Names
Reductio log absurdum.
GetParameterValues is generally not interesting from the
device, as the ACS will have the values the log could supply.
Only log the first couple parameters, and suppress the rest.

GetParameterNamesResponse is similarly not very interesting
in terms of debugging real problems.

One step toward slowing rate of growth in b/20252550

Change-Id: Ia9f9fb47c7022e1a06ba97fba59467044ec45590