Add ForceTvBoxWifiClient experiment.
The ForceTvBoxWifiClient experiment tries, currently with minimal
intelligence, to join the home WLAN. It will be used in conjunction with
the (now renamed) WhatIfTV experiment to test the performance of
wireless TV in a home.

This includes a supporting test change: the mock at
dm/testdata/binwifi/binwifi now appends its output file rather than
overwriting it, so all calls made to /bin/wifi in a single method
invocation can be examined in tests.

Change-Id: I72c6ec089bb78189d941defc60019025365d3651