Helpers to ensure circular references aren't created in unit tests.
pyinotify was using ctypes, which seems to create some unavoidable circular
references that need to be cleaned up by the garbage collector, but
mercifully that doesn't need to happen too often, so we explicitly ignore
that source of garbage objects.  With that cwmptypes_test and
filenotifier_test pass cleanly.  I haven't tried the other tests yet (but
there are likely to still be several major culprits).

Change-Id: I9d0a025961c01cf999f6b667fa5c2519b3cc4009