catawampus: checks previous ginstall completion.
In the event that cwmpd is killed or crashes during a ginstall
installation, and the installation completes independent of cwmp, there
will now be a check for the status file ginstall leaves behind.  If it
exists, installation will continue as per usual, without starting a new
instance of ginstall.

b/22850099

---

Original change was reverted due to test breakage.  This includes fixed
tests to prevent breakage and ensure proper test functionality.

Change-Id: If669f89c03aabadaee1e56fd2232abbcc688c6a7