catawampus: checks previous ginstall completion.
In the event that cwmpd is killed or crashes during a ginstall
installation, and the installation completes independent of cwmp, there
will now be a check for the status file ginstall leaves behind.  If it
exists, installation will continue as per usual, without starting a new
instance of ginstall.

b/22850099

Change-Id: Iec903c6cdbf97879eb67a27b354dcd7e2e45d82f