Get rid of tr.core.Exporter.Close() functions.
These were only ever used in fishy situations: where we wanted to
notify the owner object that a child was being deleted.  This resulted
in child objects that needed to know about parent objects, and it was
easy to make mistakes (for example, adding/deleting objects without
using tr.core.Handle would usually forget to call the callback).

Instead, in the few places where this was needed, just derive a new
object from dict instead, and have it call the callback on every
setitem/delitem.  As a bonus, I'm pretty sure this fixes a few edge
cases in nat.py and dnsmasq.py.

Change-Id: If9cd58090d9f52a483e43388edaa2a65aeb152c1