Store tr.cwmptypes as individual attrs instead of __Attrs dict.
We were using a per-instance __Attrs dict, but dictionaries are quite
expensive.  Instead, store each attr's value in a property named _Whatever
(for a given attr whose public name is Whatever).  This allows the
containing object to put the values in its __slots__ and thus avoid the
big dictionary.

On x86, reduces memory usage for the ~1000 periodicstats on GFRG200
from 1420k to 1152k.

Change-Id: I21cc080a91aee4037531b301344c8634eb99e2bf