periodic_statistics: merge and intern() some never-changing settings.
Several Parameter settings are almost never changed, and even when they do
change, it's to one of a very small number of values.  Combine them into a
single tuple, stringify the tuple, and intern() it so that it can be shared
across all Parameter instances.

This lets us reduce the number of __slots__ in each Parameter by about half.
While we're here, remove the __weakref__ slot, which prevents anyone from
taking a weakref to a Parameter object, but that was only ever used in the
unit test.  Since there are no longer any circular references, that
particular test is much less tricky, so just remove it.

On x86, reduces memory usage for the ~1000 periodicstats on GFRG200
from 1152k to 600k.

Change-Id: I1f45bd20700ea67104550fb3439be605802f0b92