Fix circular references in tr.core.AutoDict and tr.session.cache.
Most users of AutoDict would pass self.somefunc into it as callbacks,
and then keep a ref to the AutoDict in self.  To fix it, make AutoDict
be careful to hold only weak refs to member functions (well, actually,
weakrefs to the 'self' object and the class version of the member
function), thus breaking the loop.

This led to a problem with tr.session.cache, which wraps member
functions in a way that they couldn't be weakref'd by the above
AutoDict change.  Adjust tr.session.cache so that it, too, holds only a
weak reference to its inner function's object (if any).

Add some tests to make sure the circular references stay fixed.

Change-Id: Ia87273f7ce618c92482c87b27c0c251c6f2940e5