subprocess.py: fix memory leak when Popen() fails at creation time.
This one is really tricky.  The original code:
- saves the exception+traceback as local variables.
- then it does some cleanup, which might throw an exception and thus
  lose the old traceback.
- then it re-throws the old traceback.

Unfortunately, this creates a circular object reference, because the
traceback itself contains the current stack frame (since it's a
traceback after all!) and now the current stack frame contains the
traceback!  The stack frame contains self, ie.  the Popen object, and
the Popen object has a destructor.  python doesn't know how to properly
clean up circular references when they have destructors (since there's
no safe way to break the circle), hence the memory leak.

We fix it by making it even more complex: re-throw the original
exception+traceback, but do it inside a try/finally.  In the finally
clause, del the traceback from the local stack frame, thus breaking the
loop.

No tests in this commit, but the following one adds additional tests
that fail when this bug is present.

Change-Id: I13a498156125751e9abb65c039efe40eae4d7c8d