techui: remove Wifi APs, isostream, codewords.
The Wifi APs tab isn't very useful in its current form.
Remove it from the Tech UI until we're happy with it.

The isostream tab doesn't currently work. Supply one fix
by setting the expected bandwidth, but hide the tab until
we are happy with it.

The corrected and uncorrected MoCA codeword graphs always show
zero. Not sure if they are broken, or if those counters are just
not as useful as we thought. Remove those graphs.

The Wifi Clients Estimated Throughput (aka wifiblaster_graph) is
always empty, remove it until we have it working.

Change-Id: Ic0e955b294976b14f20f5adbfd3aaa2cd98835f7