Always export wifi policy; never run /bin/wifi from cwmpd.
The exact command to run /bin/wifi and start an AP is now exported to
a file per-interface, whether or not is-wireless-client is true.  A
separate file will be present iff the access point should actually be
run.

We decided to run a small shell script on all devices which are not
capable of being wireless clients, e.g. Windcharger.  This script will
mirror the AP- management functionality of conman.  The main results
of this are:

- Smaller cwmpd codebase and test surface.
- cwmpd no longer blocks waiting for /bin/wifi to start hostapd or
  wpa_supplicant.

A side effect is that WhatIfWifi will not run on devices for which
is- wireless- client is false, but this should not be a problem since
that should be exactly the set of devices which are not candidates for
wireless TV.  If at some point we do want to explore wireless TV on
e.g. Windcharger, we will have to revisit this.

Change-Id: Ia43893405c5785cb25099d31b070ac0ff3eb977d