hosts: remove SSDP and ASUS
At this point we're never going to make use of the
ASUS & SSDP information in Device.Hosts. Additionally,
SSDP is currently implemented with a blocking call to
ssdp_poll, a pattern which we're gradually eliminating
from catawampus.

Remove the SSDP and ASUS functionality, we'll report
taxonomy information using these protocols via a different
channel.

Change-Id: I93e64be31937a114999944765c234d74d512e75c