Better handling of KeyError() in AddObjects/SetParameterValues.
We would often only report the name of the particular path element that
couldn't be found, rather than the full path of that element.  So when
looking up This.That.1.Whatever, we might just raise KeyError('1').  A lot
of things might have a '1' sub-element, making it hard to debug.  Now, we'd
raise KeyError('This.That.1') to indicate that '1' is the first element that
couldn't be found, inside This.That.

We also were incorrectly throwing a full cwmp "invalid arguments" fault for
objects that couldn't be found; the correct thing is to return a
SetParameterValues fault and enumerate the exact parameters that can't be
found, in a particular format, like we do when we're having trouble setting
values.  Unfortunately, it's inefficient to look up values after the first
failure, so we still only return the first unknown name, but that's still an
improvement.

Change-Id: Id773ff29b57c57078ce7e42f22877385b0c24899