Export wifi policy even when RadioEnabled is false.
RadioEnabled actually means that the access point should be enabled;
prior to wireless CPE there was no distinction.  RadioEnabled=false
does NOT mean that the device should not attempt to operate as
wireless CPE, which is what was happening.

Instead, don't export wifi policy when it's empty, and make it empty
when there is no SSID.

BUG=30169106

Change-Id: I6c08184c20f0780e479440bed2fb0f47ad7f355f