DiagUI and TechUI: cull unneeded fields and data.
The DiagUI list of wireless clients lists them by MAC address and
whether they are authenticated. This isn't very useful, and the
TechUI has a better basis to start from as it includes the Hostname
and nicer formatting. Remove it from the DiagUI, we'll focus on the
TechUI's hosts list.

The DiagUI list of wired clients only includes MoCA, not Ethernet.
Also remove this, we'll focus on the TechUI list of clients.

TechUI is populating but not using wifiblaster results, which we
want to remove from the data model. Remove the vestigial remnants
of it from the TechUI.

Remove the Username from the DiagUI. That is used for contacting
the ACS, never of possible interest to the customer, and we don't
use it for our ACS connections anyway.

Change-Id: I1a7aa4c731217074d3cdc4b2b7422cc2821e5c4d