techui: isostream fixes, reliable graphs.
Don't exit catawampus if the other end of the isostream
run stops responding, by catching the IOError in main.py.
Also stop polling the client if it dies, by checking for
connection refused to set clientRunning to false.

Tighten the callback timeout from 5 seconds to 1. This timeout
was originally intended to be the last resort, it was
expected that we'd use inotify to immediately update
graphs when new data was available. However trying to
do that was consuming 60% of the CPU at all times, even
when no-one was looking at the TechUI. That isn't
supportable, and we now rely on the timeout much more
often.

Make the Javascript implement a timeout and retry. A single
error should not make the page stop working.

Change-Id: Icc77e09aeacac1006e81089dee686fb708804244