dnsmasq: only write config if we have one.
catawampus starts up with dnsmasq.py having an empty ACS URL:
DhcpAcsUrl = ['']

On the very first Inform cpe_management_server passes back the ACS URL
it used, and this is treated as a change which should rewrite
dnsmasq.conf.

However the rest of the DHCP config has not been supplied at that point,
so it writes out a mostly empty dnsmasq.conf with no pools
configured. It *does* have an ACS URL, so we’ve got that going
for us, which is nice, but it breaks DHCP until the ACS pushes
the full config... which won't happen if the WAN is down.

Instead, only write a dnsmasq.conf if we actually have a DHCP
configuration from the ACS, defined by at least one Pool having
been explicitly enabled or disabled.

Change-Id: I3c079f241de5d16661dc8b77777dec201c5d54f2