set-acs: correctly handle reverting to a previously-blessed URL.
If you ran this sequence:
   set-acs dhcp6 http://1
   set-acs bless http://1
   set-acs dhcp6 http://2
   set-acs dhcp6 http://1

It would keep trying http://2.  If http://2 was completely invalid,
eventually someone would time it out, and we'd fall back to http://1, and
all would be fine.  But if http://2 actually connected but didn't finish a
successful session, it would neither be blessed nor timed out, and this bug
prevented the DHCPv6 client from switching us back to the old blessed URL.

Also:
- Log a message when an URL is ignored because it's already timed out, just
  in case it helps with diagnostics later.
- Add set-acs.test to 'make test'.  (It also runs on the actual device
  during integration tests, which is important in case someone uses a
  bashism by accident, but there's no reason not to run it sooner
  here.)

Change-Id: I15fd638bcd3fc142853b2677b8fcc0fca19c2c9c