Fix 'bless' operation after HTTP redirect and/or noautoprov flags.
In b/32680701, it turned out that we were blessing the wrong URL after an
HTTP redirect, so that the blessing of the original URL never takes effect.
That isn't super serious except that we might sometimes fall back to the
previously-blessed URL, whatever it was, in the event that the URL changed.

However, upon investigating that I found out that recent changes to add
the ?options=noautoprov flag were also interfering with our ability to bless
and time out URLs.

To fix both problems, let's move the code around so that the "ACS URL" is a
different thing than the "original ACS URL" and the "qualified ACS URL," and
track them all separately.

Also, let's add some tests to make sure we bless the right URLs at the right
time.

Change-Id: I01f03d1f5f9693e78fc6b5db38e67c6f0722493e